버그 수정 및 채팅 리팩토링 완료

+++ src/controllers/chat/chatRoomController.ts
@@ -1,23 +1,47 @@
+import { Types } from "mongoose";
 import * as ChatModels from "../../models/chat";
 import type { UserAttributes } from "../../models/rdbms/User";
-import type { RequestAttributes } from "../../models/rdbms/Request";
-
+import { models as RDBModels } from "../../models/rdbms";
 import logger from "../../utils/logger";
 
 const { ChatRoom, ChatUser, Unread, ChatContent } = ChatModels;
+const { Request, User } = RDBModels;
 
+// TODO: should be transactional!! -> Need new MongoDB setting
 export const createChatRoom = async (
-    request: RequestAttributes,
-    consumer: UserAttributes,
-    participants: UserAttributes[],
+    requestId: number,
+    consumerId: Buffer,
+    participantIds: Buffer[],
 ) => {
     try {
-        const chatRoomInstance = await ChatRoom.create({
-            request_id: request.request_id,
-            consumer_id: consumer.user_id,
-            participant_ids: participants.map((u) => u.user_id),
+        const request = await Request.findOne({
+            where: { request_id: requestId },
+            raw: true,
         });
 
+        const consumer = await User.findOne({
+            where: { user_id: consumerId },
+            raw: true,
+        });
+
+        const participants = await User.findAll({
+            where: { user_id: participantIds },
+            raw: true,
+        });
+
+        if (!request || !consumer || !participants) {
+            logger.error("No data");
+            return;
+        }
+
+        const chatRoomInstance = await ChatRoom.create({
+            request_id: request.request_id,
+            title: request.title,
+            consumer_id: consumerId,
+            participant_ids: participantIds,
+            message_seq: 1, // Default message will be added below
+        });
+        console.log("chatroominstance", chatRoomInstance);
         const res = await Promise.all(
             participants.map(async (parti) => {
                 await Unread.create({
@@ -30,9 +54,9 @@ export const createChatRoom = async (
         const chatContent = await ChatContent.create({
             chatroom: chatRoomInstance,
             seq: 0,
-            content_type: "alert",
+            content_type: "alarm",
             content: "방이 생성되었어요",
-            sender_id: null,
+            sender_id: Buffer.from([0]),
             image_url: "",
         });
         return chatRoomInstance;
@@ -44,8 +68,8 @@ export const createChatRoom = async (
     }
 };
 
-export const getChatRoomById = async (objectId: string) => {
-    return await ChatRoom.findById(objectId);
+export const getChatRoomById = async (chatRoomId: Types.ObjectId) => {
+    return await ChatRoom.findById(chatRoomId);
 };
 
 export const getChatRoomsByRequestId = async (requestId: number) => {
@@ -56,18 +80,30 @@ export const getAllChatRoomsByUser = async (user: UserAttributes) => {
     return await ChatRoom.find({ participant_ids: user.user_id });
 };
 
-export const getAliveChatRoomsByUser = async (user: UserAttributes) => {
-    const chatuser = await ChatUser.findOne({ user_id: user.user_id });
+export const getAliveChatRoomsByUser = async (userId: Types.ObjectId) => {
+    const chatUser = await ChatUser.findOne({ user_id: userId });
+
+    if (chatUser === null) {
+        return undefined;
+    }
 
     return await ChatRoom.find({
-        participant_ids: chatuser.user_id,
+        participant_ids: chatUser.user_id,
         // request_id of chat rooms whose requests are not done yet should be bigger than 0
         request_id: { $gte: 0 },
     });
 };
 
-export const delChatRoomsByRequest = async (request: RequestAttributes) => {
+export const delChatRoomsByRequest = async (requestId: number) => {
     // delete unread
+    const request = await Request.findOne({
+        where: { request_id: requestId },
+        raw: true,
+    });
+    if (request === null) {
+        logger.error("No such request");
+        return undefined;
+    }
     const chatRooms = await ChatRoom.find({ request_id: request.request_id });
     const chatRoomIds = chatRooms.map((chatRoom) => chatRoom._id);
     await Unread.deleteMany({ chatroom_id: { $in: chatRoomIds } });
@@ -79,10 +115,64 @@ export const delChatRoomsByRequest = async (request: RequestAttributes) => {
     );
 };
 
-export const delChatRoom = async (chatRoomId: string) => {
+export const delChatRoom = async (chatRoomId: Types.ObjectId) => {
     const chatRoom = await ChatRoom.findOne({ _id: chatRoomId });
     if (chatRoom === null) {
+        throw new Error("No such chatroom");
     }
+
+    await Unread.deleteMany({ chatroom_id: chatRoom._id });
+
+    return await ChatRoom.updateOne(
+        { _id: chatRoom._id },
+        { $set: { request_id: -1 * chatRoom.request_id } },
+    );
 };
 
-export const leaveChatRoom = async (chatRoomId: string, userId: string) => {};
+export const actionCompleteRecruit = async (
+    requestId: number,
+    consumerId: Buffer,
+    providerIds: Buffer[],
+) => {
+    // Find chatRooms that need to be deleted
+    const toDelChatRooms = await ChatRoom.find({
+        $and: [
+            { request_id: requestId },
+            { participant_ids: { $nin: providerIds } },
+        ],
+    });
+    await Promise.all(
+        toDelChatRooms.map(async (room) => {
+            return await delChatRoom(room._id);
+        }),
+    );
+
+    // Won't create unnecessary group room
+    if (providerIds.length !== 1) {
+        await createChatRoom(requestId, consumerId, [
+            ...providerIds,
+            consumerId,
+        ]);
+    }
+
+    // TODO: Push Update chatroom request
+};
+
+// Needed?
+export const leaveChatRoom = async (
+    chatRoomId: Types.ObjectId,
+    userId: Buffer,
+) => {
+    const chatRoom = await ChatRoom.findById(chatRoomId);
+
+    if (chatRoom === null) {
+        throw new Error("No such room");
+    }
+
+    const newIds = chatRoom.participant_ids.filter((id) => !id.equals(userId));
+
+    return await ChatRoom.updateOne(
+        { _id: chatRoom._id },
+        { participant_ids: newIds },
+    );
+};
